Add ProxyCard as a type.
I hope we don't have a lot more, but this is a good reduction.
